+++ src/main/kotlin/io/github/realyusufismail/armourandtoolsmod/core/init/RecipeCategoriesInit.kt
@@ -21,7 +21,10 @@ package io.github.realyusufismail.armourandtoolsmod.core.init
 import com.google.common.base.Supplier
 import com.google.common.base.Suppliers
 import com.google.common.collect.ImmutableList
+import io.github.realyusufismail.armourandtoolsmod.core.blocks.armour.CustomArmourCraftingTableRecipe
+import io.github.realyusufismail.armourandtoolsmod.core.blocks.armour.book.CustomArmourCraftingBookCategory
 import io.github.realyusufismail.armourandtoolsmod.core.blocks.tool.CustomToolCraftingTableRecipe
+import io.github.realyusufismail.armourandtoolsmod.core.blocks.tool.book.CustomToolsCraftingBookCategory
 import net.minecraft.client.RecipeBookCategories
 import net.minecraft.world.item.ItemStack
 import net.minecraft.world.item.Items
@@ -29,55 +32,42 @@ import net.minecraftforge.client.event.RegisterRecipeBookCategoriesEvent
 
 object RecipeCategoriesInit {
 
-    private val ARMOUR_CRAFTING_ARMOUR: Supplier<RecipeBookCategories> =
-        Suppliers.memoize {
-            RecipeBookCategories.create(
-                "ARMOUR_CRAFTING_ARMOUR", ItemStack(ItemInit.AQUMARINE_CHESTPLATE.get()))
-        }
-
     private val ARMOUR_CRAFTING_SEARCH: Supplier<RecipeBookCategories> =
         Suppliers.memoize {
-            RecipeBookCategories.create("ARMOUR_CRAFTING_SEARCH", ItemStack(Items.COMPASS))
+            RecipeBookCategories.create("ARMOUR_CRAFTING", ItemStack(Items.COMPASS))
         }
 
     private val ARMOUR_CRAFTING_HELMET: Supplier<RecipeBookCategories> =
         Suppliers.memoize {
-            RecipeBookCategories.create("ARMOUR_CRAFTING_HELMET", ItemStack(Items.DIAMOND_HELMET))
+            RecipeBookCategories.create(
+                "ARMOUR_CRAFTING_HELMET", ItemStack(ItemInit.AQUMARINE_HELMET.get()))
         }
 
     private val ARMOUR_CRAFTING_CHESTPLATE: Supplier<RecipeBookCategories> =
         Suppliers.memoize {
             RecipeBookCategories.create(
-                "ARMOUR_CRAFTING_CHESTPLATE", ItemStack(Items.DIAMOND_CHESTPLATE))
+                "ARMOUR_CRAFTING_CHESTPLATE", ItemStack(ItemInit.AQUMARINE_CHESTPLATE.get()))
         }
 
     private val ARMOUR_CRAFTING_LEGGINGS: Supplier<RecipeBookCategories> =
         Suppliers.memoize {
             RecipeBookCategories.create(
-                "ARMOUR_CRAFTING_LEGGINGS", ItemStack(Items.DIAMOND_LEGGINGS))
+                "ARMOUR_CRAFTING_LEGGINGS", ItemStack(ItemInit.AQUMARINE_LEGGINGS.get()))
         }
 
     private val ARMOUR_CRAFTING_BOOTS: Supplier<RecipeBookCategories> =
         Suppliers.memoize {
-            RecipeBookCategories.create("ARMOUR_CRAFTING_BOOTS", ItemStack(Items.DIAMOND_BOOTS))
+            RecipeBookCategories.create(
+                "ARMOUR_CRAFTING_BOOTS", ItemStack(ItemInit.AQUMARINE_BOOTS.get()))
         }
 
     private val ARMOUR_CRAFTING_MISC: Supplier<RecipeBookCategories> =
         Suppliers.memoize {
-            RecipeBookCategories.create("ARMOUR_CRAFTING_MISC", ItemStack(Items.SHIELD))
-        }
-
-    private val ARMOUR_CRAFTING_ALL: Supplier<RecipeBookCategories> =
-        Suppliers.memoize {
-            RecipeBookCategories.create("ARMOUR_CRAFTING_ALL", ItemStack(Items.DIAMOND_CHESTPLATE))
+            RecipeBookCategories.create(
+                "ARMOUR_CRAFTING_MISC", ItemStack(ItemInit.AQUMARINE_HELMET.get()))
         }
 
     // Tool
-    private val TOOL_CRAFTING_TOOL: Supplier<RecipeBookCategories> =
-        Suppliers.memoize {
-            RecipeBookCategories.create("TOOL_CRAFTING_TOOL", ItemStack(ItemInit.RUBY_SWORD.get()))
-        }
-
     private val TOOL_CRAFTING_SEARCH: Supplier<RecipeBookCategories> =
         Suppliers.memoize {
             RecipeBookCategories.create("TOOL_CRAFTING_SEARCH", ItemStack(Items.COMPASS))
@@ -113,28 +103,20 @@ object RecipeCategoriesInit {
             RecipeBookCategories.create("TOOL_CRAFTING_MISC", ItemStack(Items.DIAMOND_SHOVEL))
         }
 
-    private val TOOL_CRAFTING_ALL: Supplier<RecipeBookCategories> =
-        Suppliers.memoize {
-            RecipeBookCategories.create("TOOL_CRAFTING_ALL", ItemStack(Items.DIAMOND_SHOVEL))
-        }
-
     fun registerRecipeBookCategories(event: RegisterRecipeBookCategoriesEvent) {
         event.registerBookCategories(
             RecipeBookTypesInit.ARMOUR_CRAFTING,
             ImmutableList.of(
-                ARMOUR_CRAFTING_ARMOUR.get(),
                 ARMOUR_CRAFTING_SEARCH.get(),
                 ARMOUR_CRAFTING_HELMET.get(),
                 ARMOUR_CRAFTING_CHESTPLATE.get(),
                 ARMOUR_CRAFTING_LEGGINGS.get(),
                 ARMOUR_CRAFTING_BOOTS.get(),
-                ARMOUR_CRAFTING_MISC.get(),
-                ARMOUR_CRAFTING_ALL.get()))
+                ARMOUR_CRAFTING_MISC.get()))
 
         event.registerAggregateCategory(
             ARMOUR_CRAFTING_SEARCH.get(),
             ImmutableList.of(
-                ARMOUR_CRAFTING_ARMOUR.get(),
                 ARMOUR_CRAFTING_HELMET.get(),
                 ARMOUR_CRAFTING_CHESTPLATE.get(),
                 ARMOUR_CRAFTING_LEGGINGS.get(),
@@ -142,61 +124,38 @@ object RecipeCategoriesInit {
                 ARMOUR_CRAFTING_MISC.get()))
 
         event.registerAggregateCategory(
-            ARMOUR_CRAFTING_ALL.get(),
-            ImmutableList.of(
-                ARMOUR_CRAFTING_ARMOUR.get(),
-                ARMOUR_CRAFTING_HELMET.get(),
-                ARMOUR_CRAFTING_CHESTPLATE.get(),
-                ARMOUR_CRAFTING_LEGGINGS.get(),
-                ARMOUR_CRAFTING_BOOTS.get(),
-                ARMOUR_CRAFTING_MISC.get()))
+            ARMOUR_CRAFTING_HELMET.get(), listOf(ARMOUR_CRAFTING_HELMET.get()))
 
         event.registerAggregateCategory(
-            ARMOUR_CRAFTING_ARMOUR.get(),
-            ImmutableList.of(
-                ARMOUR_CRAFTING_HELMET.get(),
-                ARMOUR_CRAFTING_CHESTPLATE.get(),
-                ARMOUR_CRAFTING_LEGGINGS.get(),
-                ARMOUR_CRAFTING_BOOTS.get()))
+            ARMOUR_CRAFTING_CHESTPLATE.get(), listOf(ARMOUR_CRAFTING_CHESTPLATE.get()))
 
-        // TODO: Have a look later
-        event.registerRecipeCategoryFinder(RecipeTypeInit.ARMOUR_CRAFTING.get()) { it ->
-            if (it is CustomToolCraftingTableRecipe) {
-                return@registerRecipeCategoryFinder ARMOUR_CRAFTING_ARMOUR.get()
+        event.registerAggregateCategory(
+            ARMOUR_CRAFTING_LEGGINGS.get(), listOf(ARMOUR_CRAFTING_LEGGINGS.get()))
+
+        event.registerAggregateCategory(
+            ARMOUR_CRAFTING_BOOTS.get(), listOf(ARMOUR_CRAFTING_BOOTS.get()))
+
+        event.registerRecipeCategoryFinder(RecipeTypeInit.ARMOUR_CRAFTING.get()) {
+            if (it is CustomArmourCraftingTableRecipe) {
+                when (it.category()) {
+                    CustomArmourCraftingBookCategory.ARMOUR_HEAD -> ARMOUR_CRAFTING_HELMET.get()
+                    CustomArmourCraftingBookCategory.ARMOUR_CHEST ->
+                        ARMOUR_CRAFTING_CHESTPLATE.get()
+                    CustomArmourCraftingBookCategory.ARMOUR_LEGS -> ARMOUR_CRAFTING_LEGGINGS.get()
+                    CustomArmourCraftingBookCategory.ARMOUR_FEET -> ARMOUR_CRAFTING_BOOTS.get()
+                    else -> {
+                        ARMOUR_CRAFTING_MISC.get()
+                    }
+                }
             } else {
-                return@registerRecipeCategoryFinder ARMOUR_CRAFTING_MISC.get()
+                ARMOUR_CRAFTING_MISC.get()
             }
         }
 
         event.registerBookCategories(
             RecipeBookTypesInit.TOOL_CRAFTING,
             ImmutableList.of(
-                TOOL_CRAFTING_TOOL.get(),
                 TOOL_CRAFTING_SEARCH.get(),
-                TOOL_CRAFTING_SWORD.get(),
-                TOOL_CRAFTING_PICKAXE.get(),
-                TOOL_CRAFTING_AXE.get(),
-                TOOL_CRAFTING_SHOVEL.get(),
-                TOOL_CRAFTING_HOE.get(),
-                TOOL_CRAFTING_MISC.get(),
-                TOOL_CRAFTING_ALL.get()))
-
-        event.registerAggregateCategory(
-            TOOL_CRAFTING_SEARCH.get(),
-            ImmutableList.of(
-                TOOL_CRAFTING_TOOL.get(),
-                TOOL_CRAFTING_SWORD.get(),
-                TOOL_CRAFTING_PICKAXE.get(),
-                TOOL_CRAFTING_AXE.get(),
-                TOOL_CRAFTING_SHOVEL.get(),
-                TOOL_CRAFTING_HOE.get(),
-                TOOL_CRAFTING_MISC.get(),
-                TOOL_CRAFTING_ALL.get()))
-
-        event.registerAggregateCategory(
-            TOOL_CRAFTING_ALL.get(),
-            ImmutableList.of(
-                TOOL_CRAFTING_TOOL.get(),
                 TOOL_CRAFTING_SWORD.get(),
                 TOOL_CRAFTING_PICKAXE.get(),
                 TOOL_CRAFTING_AXE.get(),
@@ -204,11 +163,43 @@ object RecipeCategoriesInit {
                 TOOL_CRAFTING_HOE.get(),
                 TOOL_CRAFTING_MISC.get()))
 
+        event.registerAggregateCategory(
+            TOOL_CRAFTING_SEARCH.get(),
+            ImmutableList.of(
+                TOOL_CRAFTING_SWORD.get(),
+                TOOL_CRAFTING_PICKAXE.get(),
+                TOOL_CRAFTING_AXE.get(),
+                TOOL_CRAFTING_SHOVEL.get(),
+                TOOL_CRAFTING_HOE.get(),
+                TOOL_CRAFTING_MISC.get()))
+
+        event.registerAggregateCategory(
+            TOOL_CRAFTING_SWORD.get(), listOf(TOOL_CRAFTING_SWORD.get()))
+
+        event.registerAggregateCategory(
+            TOOL_CRAFTING_PICKAXE.get(), listOf(TOOL_CRAFTING_PICKAXE.get()))
+
+        event.registerAggregateCategory(TOOL_CRAFTING_AXE.get(), listOf(TOOL_CRAFTING_AXE.get()))
+
+        event.registerAggregateCategory(
+            TOOL_CRAFTING_SHOVEL.get(), listOf(TOOL_CRAFTING_SHOVEL.get()))
+
+        event.registerAggregateCategory(TOOL_CRAFTING_HOE.get(), listOf(TOOL_CRAFTING_HOE.get()))
+
+        event.registerAggregateCategory(TOOL_CRAFTING_MISC.get(), listOf(TOOL_CRAFTING_MISC.get()))
+
         event.registerRecipeCategoryFinder(RecipeTypeInit.TOOL_CRAFTING.get()) {
             if (it is CustomToolCraftingTableRecipe) {
-                return@registerRecipeCategoryFinder TOOL_CRAFTING_TOOL.get()
+                when (it.category()) {
+                    CustomToolsCraftingBookCategory.SWORD -> TOOL_CRAFTING_SWORD.get()
+                    CustomToolsCraftingBookCategory.PICKAXE -> TOOL_CRAFTING_PICKAXE.get()
+                    CustomToolsCraftingBookCategory.AXE -> TOOL_CRAFTING_AXE.get()
+                    CustomToolsCraftingBookCategory.SHOVEL -> TOOL_CRAFTING_SHOVEL.get()
+                    CustomToolsCraftingBookCategory.HOE -> TOOL_CRAFTING_HOE.get()
+                    CustomToolsCraftingBookCategory.MISC -> TOOL_CRAFTING_MISC.get()
+                }
             } else {
-                return@registerRecipeCategoryFinder TOOL_CRAFTING_MISC.get()
+                TOOL_CRAFTING_MISC.get()
             }
         }
     }